+++ apps/api/database/migrations/2024_05_16_144146_create_orders_table.php
@@ -31,13 +31,13 @@ return new class extends Migration
             $table->longText('description')
                 ->nullable();
 
-            $table->float('service_fee', 8, 2)
+            $table->decimal('service_fee', 10, 2)
                 ->nullable();
 
-            $table->float('total', 8, 2)
+            $table->decimal('total', 10, 2)
                 ->nullable();
 
-            $table->float('discount', 8, 2)
+            $table->decimal('discount', 10, 2)
                 ->nullable();
 
             $table->softDeletes();
